Rework upsert individual function
To prevent gaps in the `individual_id` sequence and unnecessary updates to records
in `warehouse.individual` when there are no changes, the `upsert_individual` function
is being updated to first check if a record exists and then only perform the update
if the relevant columns have different data. Preventing unnecessary updates will help
reduce bloat and provide a more accurate modified date on the record.