Add index to identifier.barcode column
To improve performace of sort, filters, and joins that use identifier.barcode column, an
index is being added. One of the drivers for this change is to address recurring issue of
`shipping.uw_reopening_ehs_reporting_v1` view hanging indefinitely, which includes a join
with `shipping.return_results_v3` on the barcode column.