Warn when sample identifiers are changed during upsert
upsert_sample can allow identifiers to change for a given sample.
It should be rare that a sample identifier or collection identifier
is changed from one non-null value to another. When this happens,
log a warning so that we can confirm with the data source (typically,
the lab) that this change was intended. And if so, we can do any
manual cleanup as necessary for participants results or linelist
reporting, for example.