Prevent Updates to Unchanged Sample Records
This change checks for identifier and metadata changes in sample records prior to upserting
them into the warehouse table. If no changes are detected we can safely skip updates and
thereby make less requests to our table. Note we check identifiers separately because we
often upsert with one or the other.

This change also adds a warning if the function is called and requested to update identifiers
but is only passed 1 identifier. This would be useful if we have a sample with both identifiers
filled out and we are updating it with only one identifier passed in (this would erase the
existing identifier).

Co-Authored-By: Kristen Schwabe-Fry 2600836+kschwabefry@users.noreply.github.com