Add `mint_identifiers` function to improve performance
A `mint_identifiers` function is being added to improve performance of identifier
minting by handling more of the process in postgres and less in python. The function
will generate uuids and check the hamming distance of the last 8 characters against
all existing barcodes. It will continue trying new identifiers until the distance check
passes, at which point it will insert a record into the identifier table. The function will repeat
this process until it has reached the indicated `number_to_mint` and returns the records that
were inserted.