Allow Whitespace to be Ignored when Loading NDJSON Files
This change allows whitespace to be ignored when loading NDJSON files by adding
an optional flag `ignore_empty_lines = True` to the function definition. This
is in response to a Pandas change that adds a (superfluous) newline to the end
of NDJSON dumps. This will allow our loading code to handle either behavior
based on what we want. Note NDJSON specification [here](https://github.com/ndjson/ndjson-spec).

This change also adds additional doctests for this altered behavior and fixes
broken tests as a result of the Pandas changes.