Add a Case Insensitive Option to Manifest Diff
Adds an option `--ignore-case` to the manifest diff ID3C command. Running
the command without this option retains the current behavior, which is to
do a deep diff of manifest files while remaining sensitive to string case.
Running the command with this option will perform the same diff, but will
treat upper and lowercase strings the same.