Add Decorator for ID3C CLI Jobs that Use a REDCap Project
Adds a function decorator modeled after the `with_database_session` decorator
that can be used to provide a REDCap project to a CLI routine. Reduces code
duplication across a couple of our jobs by folding in the `api-url`, `project-id`,
and `token` options into a decorator and using them to construct a project. This
project can then be passed directly to the invoked command.