Suppress a MyPy Type Error in ID3C Url Library
A new version of pytest and mypy causes a type regression in our URL module. It
seems like it complains we are potentially returning a NoneType object, but
we should be ok since we are returning an object from Urllibs URL classes. Suppressing
this error should be fine for now.