Include `access_role` value in genomic sequencing ETL
An `access_role` column was recently added to the `sample`, `sequencing_read_set`, `genomic_sequence`, and `consensus_genome` tables of the `warehouse` schema to allow data with restricted access to be included in these tables. These changes include `access_role` in the appropriate class definitions, and updates ETLs to propagate the `access_role` value from a `sample` record to the related sequencing records.

The sample record will continue server as the primary `access_role` value, but must also be stored in each table that uses row level security.